refactor cmd structure
