--- bin/deux-list
@@ -1,15 +1,0 @@
-#!/usr/bin/env node
-
-require('babel-register')
-const program = require('commander')
-const cli = require('../src/cli')
-
-program.parse(process.argv)
-
-cli()
-  .then(({db, list}) => {
-    list(db, program.args)
-  })
-  .catch(() => {
-    process.exit(0)
-  })
