+++ package.json
@@ -44,6 +44,7 @@
     "semver": "^5.3.0",
     "string": "^3.3.3",
     "validate.js": "^0.11.1",
+    "weft": "^1.1.0",
     "wp-get-file-header": "^0.2.0",
     "wp-plugin-search": "^1.1.0"
   },
