+++ src/cli/add-cmd/component.js
@@ -3,12 +3,12 @@ import {existsSync} from 'fs'
 import inquirer from 'inquirer'
 import faker from 'faker'
 import _s from 'string'
-import validator from '../../lib/validator'
-import {error, done, colorlog} from '../../lib/logger'
-import {templateDir, wpThemeDir} from '../../lib/const'
-import {compileFile} from '../../lib/utils'
-import {dbErrorHandler, getCurrentTheme, saveConfig} from '../../lib/db-utils'
-import * as message from '../../lib/messages'
+import validator from '../lib/validator'
+import {error, done, colorlog} from '../lib/logger'
+import {templateDir, wpThemeDir} from '../lib/const'
+import {compileFile} from '../lib/utils'
+import {dbErrorHandler, getCurrentTheme, saveConfig} from '../lib/db-utils'
+import * as message from '../lib/messages'
 
 export default db => {
   colorlog('Add a {New Component}')
