+++ src/cli/add-cmd/hook.js
@@ -1,4 +1,4 @@
-import {colorlog} from '../../lib/logger'
+import {colorlog} from '../lib/logger'
 
 export default () => {
   colorlog('Add {action/filter} hook')
