+++ src/cli/add-cmd/index.js
@@ -1,14 +1,14 @@
 import inquirer from 'inquirer'
-import * as message from '../../lib/messages'
-import {error, colorlog} from '../../lib/logger'
-import {validAddCommand} from '../../lib/const'
-import {capitalize} from '../../lib/utils'
-// import addHook from './add-hook'
-import addAsset from './add-asset'
-import addPlugin from './add-plugin'
-// import addFeature from './add-feature'
-import addTemplate from './add-template'
-import addComponent from './add-component'
+import * as message from '../lib/messages'
+import {error, colorlog} from '../lib/logger'
+import {capitalize} from '../lib/utils'
+import {validAddCommand} from '../lib/const'
+import addHook from './hook'
+import addAsset from './asset'
+import addPlugin from './plugin'
+import addFeature from './feature'
+import addTemplate from './template'
+import addComponent from './component'
 
 const displayPrompt = (db, cmd) => {
   switch (cmd) {
