+++ src/cli/lib/logger.js
