+++ src/cli/lib/validator.js
